requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
Babel==2.8.0
black==19.10b0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
Click==7.0
cryptography==3.3.1
cssselect==1.1.0
cssutils==1.0.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.41.0
django-appconf==1.0.4
django-ckeditor==6.0.0
django-colorfield==0.3.2
django-cors-headers==3.2.1
django-currentuser==0.5.2
django-email-extras==0.3.4
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-location-field==2.1.0
django-modeltranslation==0.16.1
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-simple-mail==2.3.3
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
html2text==2020.1.16
httplib2==0.18.1
idna==2.8
lxml==4.6.2
Markdown==3.3.3
oauth2client==4.1.3
oauthlib==3.1.0
pathspec==0.7.0
phonenumbers==8.11.4
pilkit==2.0
Pillow==8.0.1
pip==20.3.3
premailer==3.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python-gnupg==0.4.6
python3-openid==3.1.0
pytz==2019.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
setuptools==50.3.2
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wheel==0.35.1
